Extract createGame and its route out of submissions.ts
`createGame` is the whole creation path — validate, per-IP limit, moderate,
circuit-breaker, availability, quota, slug mint and claim, brief, dispatch —
and it is already the shared implementation behind two transports: the HTTP
route here and the MCP `create_game` tool, which takes it through the routes
handle. Moving it to `creation/create-game.ts` puts that shared sequence in a
file named after it rather than in the middle of a 2331-line module.

  createGameCreator(deps)        the sequence, returning a result not a reply
  registerCreateGameRoute(app)   POST /api/submissions, mapping it to statuses

`buildDispatchIssueBody` moves too, and is exported from the new file. It was
exported from submissions.ts with a comment claiming the dispatch reaper used
it; nothing outside submissions.ts imports either it or `reconstructDispatchSpec`,
which stays and now imports it. `CreateSubmissionRequestSchema` has no other
reader, so it moves with the only code that parses against it.

Two changes to the bodies, both mechanical:

  - the inline return union becomes the named `CreateGameResult`, same members
  - one `app.log` becomes the `log` dep, typed `FastifyBaseLogger` because
    `logModerationRejection` needs the full logger

Verified by diffing both against master with comments and blank lines stripped:
those two lines in `createGame`, and the route body byte-identical at 26 code
lines.

submissions.ts: 2331 -> 2103 lines, ceiling resealed. The moderation-metrics
guard test drops submissions.ts and names create-game.ts: after this the god
module moderates nothing, because it no longer holds a route that takes creator
text.
